AlgoWeirdCalculator/calculator/pyproject.toml
[tool.poetry]
name = "calculator"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["SJ <sarajanedeveloper@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pip-audit = "*"
black = {extras = ["d"], version = "*"}
ruff = "0.0.252"
mypy = "*"
pytest = "*"
pytest-cov = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "ANN", "UP", "N", "C", "B", "A", "YTT", "W", "FBT", "Q", "RUF", "I"]
ignore = ["ANN101", "C901", "F403", "F405"]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Assume Python 3.10.
target-version = "py310"

[tool.ruff.flake8-annotations]
allow-star-arg-any = true
suppress-none-returning = true


[tool.pytest.ini_options]
pythonpath = [".", "tests"]


[tool.mypy]
files = "./"
python_version = "3.10"
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
allow_untyped_defs = false
strict_equality = true

